requirements.txt
google-cloud-storage==1.22.0
google-cloud-translate==2.0.0